Trying to move to master-worker arch
