+++ worker/Dockerfile
@@ -7,8 +7,12 @@ RUN tar -xf libweb.tar.gz libwebp-0.4.3-rc1-linux-x86-64/bin/webpmux
 RUN cp libwebp-0.4.3-rc1-linux-x86-64/bin/cwebp /usr/bin
 RUN cp libwebp-0.4.3-rc1-linux-x86-64/bin/webpmux /usr/bin
 RUN rm -rf libwebp-0.4.3-rc1-linux-x86-64/ libweb.tar.gz
+# Add docker-compose-wait tool -------------------
+ENV WAIT_VERSION 2.7.2
+ADD https://github.com/ufoscout/docker-compose-wait/releases/download/$WAIT_VERSION/wait /wait
+RUN chmod +x /wait
 COPY go.mod main.go ./
-ADD handler ./handler
+ADD convert ./convert
 ADD metadata ./metadata
 RUN go mod tidy
-ENTRYPOINT ["go", "run", "main.go", "-log-level", "DEBUG"]
+ENTRYPOINT ["go", "run", "main.go"]
