+++ worker/go.mod
@@ -1,0 +1,16 @@
+module github.com/deven96/whatsticker
+
+go 1.17
+
+require github.com/adjust/rmq/v4 v4.0.5
+
+require (
+	github.com/cespare/xxhash/v2 v2.1.1 // indirect
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/go-redis/redis/v8 v8.3.2 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.6.1 // indirect
+	go.opentelemetry.io/otel v0.13.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+)
