+++ worker/metadata/raw.exif
